+++ src/main/java/com/akshitbaunthy/restapiassignmentsolution/service/EmployeeDeleteServiceImpl.java
@@ -1,27 +1,38 @@
 package com.akshitbaunthy.restapiassignmentsolution.service;
 
+import com.akshitbaunthy.restapiassignmentsolution.entity.Employee;
 import com.akshitbaunthy.restapiassignmentsolution.repository.EmployeeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.util.Optional;
 
 @Service
 public class EmployeeDeleteServiceImpl implements EmployeeDeleteService {
 
     EmployeeRepository deleteRepository;
+    EmployeeRepository readRepository;
 
     @Autowired
-    private void EmployeeDeleteServiceImpl(EmployeeRepository employeeRepository) {
+    private EmployeeDeleteServiceImpl(EmployeeRepository employeeRepository) {
         this.deleteRepository = employeeRepository;
+        this.readRepository = employeeRepository;
     }
 
     @Override
-    public String deleteById(Long id) {
-        deleteRepository.deleteById(id);
-        return "The Employee with ID " + id + " is no longer with the company.";
+    public boolean deleteById(Integer id) {
+        Optional<Employee> fetchedEmployee = readRepository.findById(id);
+        if(fetchedEmployee.isPresent()) {
+            deleteRepository.deleteById(id);
+            return true;
+        }
+        else {
+            return false;
+        }
     }
 
     @Override
-    public String deleteAll() {
-        return "The company now has no employees";
+    public void deleteAll() {
+        deleteRepository.deleteAll();
     }
 }
